deleted screens folder and shifted to components
