+++ frontend/App.js
@@ -1,38 +1,16 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import React, { useState } from "react";
-import {
-  StyleSheet,
-  Text,
-  View,
-  TouchableOpacity,
-  Modal,
-  Button,
-} from "react-native";
-import SignUpScreen from './screens/SignUpScreen';
-import LoginScreen from './screens/LoginScreen';
-import HomeScreen from './screens/HomeScreen';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import React, { useState }  from 'react';
+import { StyleSheet, View, TouchableOpacity, Modal, Button, Image, Text } from 'react-native';
+import LoginScreen from './components/LoginScreen';
+import HomeScreen from './components/HomeScreen';
 import Icon from 'react-native-vector-icons/Ionicons';
 import QRCode from 'react-native-qrcode-svg';
-import StoresScreen from './screens/StoresScreen';
-import RewardsScreen from './screens/RewardsScreen';
-import AccountScreen from './screens/AccountScreen';
-
-// const Stack = createNativeStackNavigator();
-
-// const App = () => {
-//   return (
-//     <NavigationContainer>
-//       <Stack.Navigator initialRouteName="SignUp">
-//         <Stack.Screen name="Sign Up" component={SignUpScreen} />
-//         <Stack.Screen name="Login" component={LoginScreen} />
-//         <Stack.Screen name="Home" component={HomeScreen} options={{ headerShown: false }} />
-//       </Stack.Navigator>
-//     </NavigationContainer>
-//   );
-// };
-// export default App;
+import StoresScreen from './components/StoresScreen';
+import RewardsScreen from './components/RewardsScreen';
+import AccountScreen from './components/AccountScreen';
+import RegisterScreen from './components/RegisterScreen';
 
 const Stack = createNativeStackNavigator();
 const Tab = createBottomTabNavigator();
@@ -84,10 +62,10 @@ const CustomTabBar = ({ state, descriptors, navigation }) => {
           >
             {route.name === 'QRButton' ? (
               <View style={styles.qrButton}>
-                <Icon name="qr-code" color="#fff" size={40} />
+                <Icon name="qr-code" color="#fff" size={30} />
               </View>
             ) : (
-              <Icon name={iconName} color={isFocused ? '#88C34A' : '#222'} size={25} />
+              <Icon name={iconName} color={isFocused ? '#88C34A' : '#222'} size={30} />
             )}
           </TouchableOpacity>
         );
@@ -100,8 +78,12 @@ const CustomTabBar = ({ state, descriptors, navigation }) => {
       >
         <View style={styles.modalContainer}>
           <View style={styles.modalContent}>
-            <QRCode value="https://example.com" size={200} />
-            <Button title="Close" onPress={() => setModalVisible(false)} />
+            <TouchableOpacity style={styles.closeButton} onPress={() => setModalVisible(false)}>
+              <Image source={require('./assets/close-icon.png')} style={styles.closeIcon} />
+            </TouchableOpacity>
+            <Text style={styles.modalText}>My Pouch QR</Text>
+            <QRCode value="https://www.youtube.com/watch?v=BbeeuzU5Qc8" size={200} />
+            <Text style={styles.modalText}>P 128 039</Text>
           </View>
         </View>
       </Modal>
@@ -142,10 +124,10 @@ const MainTabNavigator = () => (
 export default function App() {
   return (
     <NavigationContainer>
-      <Stack.Navigator initialRouteName="SignUp">
+      <Stack.Navigator initialRouteName="Register">
         <Stack.Screen
-          name="SignUp"
-          component={SignUpScreen}
+          name="Register"
+          component={RegisterScreen}
           options={{ headerShown: false }}
         />
         <Stack.Screen
@@ -166,7 +148,7 @@ export default function App() {
 const styles = StyleSheet.create({
   tabBarContainer: {
     flexDirection: 'row',
-    height: 75,
+    height: 90,
     borderTopWidth: 1,
     borderTopColor: '#ccc',
     backgroundColor: '#fff',
@@ -180,11 +162,11 @@ const styles = StyleSheet.create({
   },
   qrButton: {
     position: 'absolute',
-    bottom: -10,
+    bottom: 5,
     alignSelf: 'center',
     backgroundColor: '#88C34A',
-    padding: 20,
-    borderRadius: 100,
+    padding: 25,
+    borderRadius: 50,
     zIndex: 10,
   },
   modalContainer: {
@@ -194,10 +176,27 @@ const styles = StyleSheet.create({
     backgroundColor: 'rgba(0,0,0,0.5)',
   },
   modalContent: {
-    width: 300,
+    width: 280,
+    height: 400,
     padding: 20,
     backgroundColor: 'white',
     borderRadius: 10,
     alignItems: 'center',
+  },
+  modalText: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    marginBottom: 30,
+    marginTop: 30,
+  },
+  closeButton: {
+    position: 'absolute',
+    top: 10,
+    left: 10,
+    zIndex: 1, // Ensure the close button is above other content
+  },
+  closeIcon: {
+    width: 30,
+    height: 30,
   },
 });
