--- frontend/screens/AccountScreen.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
-
-const AccountScreen = () => (
-  <View style={styles.container}>
-    <Text>Account Screen</Text>
-  </View>
-);
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
-
-export default AccountScreen;
